% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertAreas.R
\docType{data}
\name{alertAreas}
\alias{alertAreas}
\title{Alert Areas Used by the National Weather Service}
\format{A \code{\link[sp:SpatialPolygons-class]{SpatialPolygons}} object of
  length 7395 whose names are 2-letter state abbreviations and 6-character
  Universal Geographic Code (\acronym{UGC}) county and zone codes. Polygons
  are specified in WGS84 coordinates.}
\source{
\url{http://www.nws.noaa.gov/geodata/}
}
\usage{
alertAreas
}
\description{
Polygons defining the states, counties, and zones used by the United States
National Weather Service (\acronym{NWS}) to define alert areas.
}
\details{
Package will be periodically updated following updates to alert
  areas by the National Weather Service.
}
\seealso{
Package \pkg{weatherAlerts}
}
\keyword{datasets}
